refactor(tests): split documentation and readiness suites out of RuntimeCoordinatorTests
The 7,279-line file covered at least eight units. The documentation
provider tests (1,168 lines) and the readiness probe/gate tests (495)
now live in their own files as extensions of the same @Suite, so the
serialized execution semantics are unchanged, and the shared fixtures
move to RuntimeCoordinatorTestSupport with target visibility instead of
being trapped file-private in the monolith.
